back/requirements.txt
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
click==8.1.3
cryptography==40.0.1
distlib==0.3.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.95.0
fastapi-cache2==0.2.1
fastapi-pagination==0.11.4
fastapi-users==10.4.1
fastapi-users-db-sqlalchemy==5.0.0
filelock==3.10.7
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
httpx-oauth==0.11.2
identify==2.5.22
idna==3.4
iniconfig==2.0.0
makefun==1.15.1
nodeenv==1.7.0
orjson==3.8.9
packaging==23.0
passlib==1.7.4
pendulum==2.1.2
platformdirs==3.2.0
pluggy==1.0.0
pre-commit==3.2.1
pycparser==2.21
pydantic==1.10.7
PyJWT==2.6.0
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
python-slugify==8.0.1
pytzdata==2020.1
PyYAML==6.0
redis==4.5.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.8
starlette==0.26.1
text-unidecode==1.3
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.21.1
virtualenv==20.21.0
